refactor: plugin management solution has confilct, change to valid/attribute package
