--- src/common/plugins/init.go
@@ -1,2 +1,0 @@
-// Package plugins initializes the plugin for extension.
-package plugins
